// apps/web/src/api/generated/model/departmentResponseDto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */

export interface DepartmentResponseDto {
  id: string;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  parentId: string | null;
  employeeCount: number;
  createdAt: string;
  updatedAt: string;
}