// apps/web/src/api/generated/model/departmentsControllerFindAllParams.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */

export type DepartmentsControllerFindAllParams = {
page?: number;
limit?: number;
search?: string;
sort?: string;
include?: string;
fields?: string;
name?: string;
};