// apps/web/src/api/generated/model/departmentsControllerFindOneParams.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */

export type DepartmentsControllerFindOneParams = {
page?: number;
limit?: number;
search?: string;
sort?: string;
include?: string;
fields?: string;
name?: string;
};